pyproject.toml
[project]
name = "pathrwkv_env"
version = "0.1.0"
requires-python = "==3.12.12"
dependencies = [
    "torch==2.9.1",
    "torchvision",
    "pytorch-lightning",
    "torchmetrics",
    "timm",
    "monai",
    "polars",
    "pyyaml",
    "numpy",
    "safetensors",
    "scikit-survival",
    "openslide-bin",
    "openslide-python",
    "pillow",
    "tqdm",
    "scipy",
    "tensorboard",
    "matplotlib",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }
